Refactor async operations on individual package results
Previously the callback for the lambda was fired before the individual results had been fully saved to s3 and emitted to SNS. This meant that results were lost from the end of the set and never returned to the client.

By refactoring to move the sending of individual results to inside the iteration over hotel ids then we can be sure that the lambda does not terminate until all results have successfully been returned to the client.